+++ CHANGELOG.md
@@ -7,7 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.44.0] - 2025-09-18
+
+### Added
+
+- `tilebox-workflows`: Added `progress_indicators` to the query response of `JobClient.find` to provide programmatic
+  access to a job's progress indicators.
+- `tilebox-workflows`: Added an `ipywidgets` based interactive display for Job objects for interactive environments like
+  Jupyter notebooks.
+
 ## [0.43.0] - 2025-09-12
+
+### Added
 
 - `tilebox-workflows`: Added progress tracking support to the `TaskRunner`.
 
@@ -255,7 +266,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released under the [MIT](https://opensource.org/license/mit) license.
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.43.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.44.0...HEAD
+[0.44.0]: https://github.com/tilebox/tilebox-python/compare/v0.43.0...v0.44.0
 [0.43.0]: https://github.com/tilebox/tilebox-python/compare/v0.42.0...v0.43.0
 [0.42.0]: https://github.com/tilebox/tilebox-python/compare/v0.41.0...v0.42.0
 [0.41.0]: https://github.com/tilebox/tilebox-python/compare/v0.40.0...v0.41.0